Move repo-connector to separate module
